Do some renaming #2584
