+++ visualization/app/codeCharta/state/effects/updateVisibleTopLabels/getNumberOfTopLabels.ts
@@ -1,10 +1,11 @@
 import { defaultAmountOfTopLabels } from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
+import { CodeMapNode } from "../../../codeCharta.model"
 
 const BUILDINGS_PER_LABEL = 100
 const MAX_NUMBER_OF_LABELS = 10
 
-export const getNumberOfTopLabels = (numberOfFiles: any) => {
-	const numberOfLabels = Math.floor(numberOfFiles.length / BUILDINGS_PER_LABEL)
+export const getNumberOfTopLabels = (codeMapNodes: CodeMapNode[]) => {
+	const numberOfLabels = Math.floor(codeMapNodes.length / BUILDINGS_PER_LABEL)
 	if (numberOfLabels <= defaultAmountOfTopLabels) {
 		return defaultAmountOfTopLabels
 	}
