+++ visualization/app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect.ts
@@ -3,7 +3,7 @@ import { Store } from "../../angular-redux/store"
 import { createEffect } from "../../angular-redux/effects/createEffect"
 import { map, withLatestFrom } from "rxjs"
 import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
-import { codeMapNodesSelector } from "../../store/appSettings/amountOfTopLabels/codeMapNodesSelector"
+import { codeMapNodesSelector } from "../../selectors/accumulatedData/codeMapNodesSelector"
 import { setAmountOfTopLabels } from "../../store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
 import { getNumberOfTopLabels } from "./getNumberOfTopLabels"
 
@@ -12,8 +12,8 @@ export class UpdateVisibleTopLabelsEffect {
 	updateVisibleTopLabels$ = createEffect(() =>
 		this.store.select(visibleFileStatesSelector).pipe(
 			withLatestFrom(this.store.select(codeMapNodesSelector)),
-			map(([, numberOfFiles]) => {
-				return setAmountOfTopLabels(getNumberOfTopLabels(numberOfFiles))
+			map(([, codeMapNodes]) => {
+				return setAmountOfTopLabels(getNumberOfTopLabels(codeMapNodes))
 			})
 		)
 	)
