+++ visualization/app/codeCharta/ui/searchPanel/matchingFilesCounter/selectors/matchingFilesCounter.selector.ts
@@ -4,7 +4,7 @@ import { searchedNodesSelector } from "../../../../state/selectors/searchedNodes
 import { blacklistSelector } from "../../../../state/store/fileSettings/blacklist/blacklist.selector"
 import { CcState } from "../../../../state/store/store"
 import { isLeaf, isPathBlacklisted } from "../../../../util/codeMapHelper"
-import { codeMapNodesSelector } from "../../../../state/store/appSettings/amountOfTopLabels/codeMapNodesSelector"
+import { codeMapNodesSelector } from "../../../../state/selectors/accumulatedData/codeMapNodesSelector"
 
 const getBlacklistedFileCount = (blacklistType: BlacklistType, nodes: CodeMapNode[], blacklist: BlacklistItem[]) =>
 	nodes.reduce((count, node) => (isPathBlacklisted(node.path, blacklist, blacklistType) ? count + 1 : count), 0)
